refactor: update the folder structure to improve code organization
