+++ src/models/mod.rs
@@ -1,4 +1,3 @@
 pub mod default;
-pub mod ide;
 pub mod jetbrains;
 pub mod vscode;
