--- src/shortcut/mod.rs
@@ -1,3 +1,0 @@
-pub mod models;
-pub mod service;
-pub mod input_device;
