--- src/shortcut/models/ide.rs
@@ -1,4 +1,0 @@
-pub enum IDE {
-    VSCode,
-    JetBrains,
-}
